chore: suggested changes Co-authored-by: Jan <66554238+vaporox@users.noreply.github.com>
