+++ packages/rest/src/lib/utils/constants.ts
@@ -15,3 +15,17 @@ export const DefaultRestOptions: Required<RESTOptions> = {
 	userAgentAppendix: `Node.js ${process.version}`,
 	version: APIVersion,
 };
+
+/**
+ * The events that the REST manager emits
+ */
+export const enum RESTEvents {
+	Debug = 'restDebug',
+	RateLimited = 'rateLimited',
+}
+
+export const ALLOWED_EXTENSIONS = ['webp', 'png', 'jpg', 'jpeg', 'gif'] as const;
+export const ALLOWED_SIZES = [16, 32, 64, 128, 256, 512, 1024, 2048, 4096] as const;
+
+export type ImageExtension = typeof ALLOWED_EXTENSIONS[number];
+export type ImageSize = typeof ALLOWED_SIZES[number];
